added permissions to index page

+++ resources/views/partials/clients/client-index.blade.php
@@ -54,9 +54,12 @@
                                         </td>
                                         <td>{{ $client->notes->first()->created_at ?? '' }}</td>
                                         <td>{{ $client->status }}</td>
-                                        <td><a href="/client/contact/{{ $client->id }}" class="btn-success btn btn-sm">Touch</a>
+                                        <td>
+                                                @can('update', $client)
+                                                    <a href="/client/contact/{{ $client->id }}" class="btn-success btn btn-sm">Touch</a>
+                                                @endcan
                                                 
-                                                <a href="/client/{{ $client->id }}/edit" class="btn btn-primary btn-sm">Edit</a>
+                                            @can('edit', $client)<a href="/client/{{ $client->id }}/edit" class="btn btn-primary btn-sm">Edit</a>@endcan
                                                 {{--  <button class="btn btn-primary"> View Notes</button>  --}}
                                                @can('delete', $client) 
                                                 <a href="/delete-client/{{ $client->id }}" id="delete" class="btn btn-danger btn-sm"> <span class="glyphicon glyphicon-remove"><strong> X </strong></span></a>
